refactor dcr_settings test
